Add handling for blocks built with the previous version
This fakes attribute migration... sort of.

It gets a bit weird because this is a dyanmic block, and because
I reused the `customTaxonomy` attribute with a different `type`.
While it worked out, in retrospect, it would have been better to
introduce a new attribute just so we wouldn't have the `customTaxonomy`
attribute registered without a type. Though, it is interesting (dangerous)
to know that it's possible to do...