Fix a handful of eslint issues
This removes the eslint config file, as the WordPress configuration
seems to cover all the bases here now.

I've opted not to worry about a few of the eslint issues - although
I could install all the `@wordpress` modules as project dependencies
and add some ignore statements in a spot or two, that doesn't quite
seem worth it.