Remove `fields => ids` as a forced query argument
In the two primary uses of WP_Query, we use have_posts() to loop
through the results and use data other than the post ID. We can
retrieve the full post object as part of the query instead.

This issue was highlighted by WordPress 6.1's introduction of
`update_post_author_caches()` and related functions. These rely
on a full post object, which they expect when in the loop.

See https://core.trac.wordpress.org/ticket/56948 for a bug report
in WordPress that will likely turn into a doing_it_wrong notice
for querying for IDs and then attempting to loop through them.

When we want to retrieve IDs only, there's no need to loop through
them. We can access the `posts` property instead for the data.

The next commit will make that change in the block's sticky post
handling.